+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
     - Show value inspector `def` on the same screen
     - Show open in editor for any form that contains line meta
+    - Remove unnamed anonymous functions from Quick Jump 
 
 ### Bugs fixed